refactor(app-opine): rename test files and move beside files under test
