+++ packages/app-opine/mod.test.js
@@ -1,8 +1,8 @@
+import { assertEquals, superdeno } from './dev_deps.js'
+
+import build from './mod.js'
+
 // TODO: Tyler. Probably better way to do this
-import { assertEquals, superdeno } from '../dev_deps.js'
-
-import build from '../mod.js'
-
 Deno.env.set('DENO_ENV', 'test')
 
 const app = build({
